feat(config): Add log level in TOML  (#14)
- Adds [log] in default toml file 


**Motivation:**
<!--
Explain here the context, and why you're making that change. What is the
problem you're trying to solve.
-->
**Modifications:**
Previously: 
Debug logs used to show up only if `--verbose` flag was provided.
**Result:**
Now, if log = "debug" is passed in TOML Or `--verbose` is passed, we
show debug logs .

**Open questions:**
<!--
(optional) Any open questions or feedback on design desired?
-->