feat(config): add default uam-permission.toml (#31)
**Motivation:**

We need to give UAM permissions for many interactions with the core
contracts.
I did not add the config in `eigen.toml` as it could have cluttered it. 


**Modifications:**

Created a separate config file called `uam-permissions.toml` , where
user can provide the permissions ,which we intend to send as
transactions when launching the devnet. For now, it has placeholder
values, we can replace them when we have the hourglass contracts
deployed.

I did not integrate it with avs create for now, since we can't exactly
test this till hourglass interaction.

**Result:**

UAM permissions

**Open questions:**

Use of a separate permissions config file instead of using the existing
`eigen.toml`