chore: add a testing section to PR template. (#53)
<!-- 
    🚨 ATTENTION! 🚨 
    
This PR template is REQUIRED. PRs not following this format will be
closed without review.
    
    Requirements:
- PR title must follow commit conventions:
https://www.conventionalcommits.org/en/v1.0.0/
- Label your PR with the correct type (e.g., 🐛 Bug, ✨ Enhancement, 🧪
Test, etc.)
    - Provide clear and specific details in each section
-->

**Motivation:**
Adding a Testing section to the PR as a reminder to actively test
changes, either manually or with automation and share those artifacts.

**Modifications:**
Improved SDLC

**Result:**
Ideally, people will be more readily reminded to share their testing
procedures and raise the testing bar on the team.

**Open questions:**