feat: refactoring hooks and standardizing uuid generation process (#78)
<!-- 
    🚨 ATTENTION! 🚨 
    
This PR template is REQUIRED. PRs not following this format will be
closed without review.
    
    Requirements:
- PR title must follow commit conventions:
https://www.conventionalcommits.org/en/v1.0.0/
- Label your PR with the correct type (e.g., 🐛 Bug, ✨ Enhancement, 🧪
Test, etc.)
    - Provide clear and specific details in each section
-->

**Motivation:**
DevKit telemetry relies on a ProjectUuid being present to uniquely
identify projects customers create. This change standardizes the
creation of a project uuid at the start of the command flow, rather than
after creation of the project. Further, the `hooks` package was not
aptly named, bloated and due for refactoring.


**Modifications:**
- Standardizes uuid generation in command middleware.
- Add error logging for metric failures.
- Refactor hooks package
- Move context package contents to common package.
- Refactor various methods for improved testing.

**Result:**
- Telemetry for `create` command now emitted.

**Testing:**
- Unit tests
- Manual testing at [PostHog](https://us.posthog.com/project/150668/sql)