docs: README updates for private preview (#75)
**Motivation:**
<!--
Explain here the context, and why you're making that change. What is the
problem you're trying to solve.
-->
Made readme edits to prepare for customer feedback.
**Modifications:**
<!--
Describe the modifications you've done from a high level. What are the
key technical decisions and why were they made?
-->
Various edits to README.md and flowchart png to prepare for customers.
**Result:**
<!--
*After your change, what will change.
-->
N/A
**Testing:**
<!--
*List testing procedures taken and useful artifacts.
-->
N/A
**Open questions:**
<!--
(optional) Any open questions or feedback on design desired?
-->
N/A

---------

Co-authored-by: Grezle <4310551+grezle@users.noreply.github.com>
Co-authored-by: JES <jamutexj@gmail.com>