fix: clear template .git before initialising new empty repo (#82)
**Motivation:**
Previously, `initGitRepo` would initialize a Git repository even if one
already existed in the target directory. This could result in corrupted
or unintended Git states. We need to ensure a clean slate every time
this function is called.

Also, if a user was to push these changes to a remote target, there is a
chance they might leak secrets in their config files. We need to make
sure that any config files are excluded from the repo.

**Modifications:**
- Added logic to detect and completely remove any existing `.git`
directory in the target directory before initializing a new Git repo.
- Added a .gitignore file to exclude config files

**Result:**
Calling `initGitRepo` now guarantees a fresh Git repository with no
remnants of any prior repo state. In addition, the potential for leaking
config now takes additional steps.