fix: avs run on devnet start unless skipped (#81)
<!-- 
    🚨 ATTENTION! 🚨 
    
This PR template is REQUIRED. PRs not following this format will be
closed without review.
    
    Requirements:
- PR title must follow commit conventions:
https://www.conventionalcommits.org/en/v1.0.0/
- Label your PR with the correct type (e.g., 🐛 Bug, ✨ Enhancement, 🧪
Test, etc.)
    - Provide clear and specific details in each section
-->

**Motivation:**
<!--
Explain here the context, and why you're making that change. What is the
problem you're trying to solve.
-->
In order to simplify user interactions with DevKit, we want as much of
the process to be up and running after `avs devnet start`.

**Modifications:**
<!--
Describe the modifications you've done from a high level. What are the
key technical decisions and why were they made?
-->
- `avs run` is called as part of `avs devnet start` unless
`--skip-deploy-contracts` or `--skip-avs-run` is flagged

**Result:**
<!--
*After your change, what will change.
-->
- reduced cognitive load on dev