chore: more informative error if fork url fails to fork a mainnet block (#101)
**Motivation:**

- It reads the cast output , and if it contains `Error: error sending
request for url`, we print put out an error saying `Please check if your
mainnet fork rpc url is up`.And we also return with error, so devnet
will not execute more if this fails.


**Modifications:**

Handle error early if checking balance or fund wallet fails , instead of
executing more code.

**Result:**


**Testing:**

The devnet unit tests were failing to get the account balances that we
fund underneath. I suspect its due to unreliable mainnet rpc. For now,
added a workaround to not fund wallets in test.

**Open questions:**
<!--
(optional) Any open questions or feedback on design desired?
-->