docs: use make install for build from source (#132)
<!-- 
    🚨 ATTENTION! 🚨 
    
This PR template is REQUIRED. PRs not following this format will be
closed without review.
    
    Requirements:
- PR title must follow commit conventions:
https://www.conventionalcommits.org/en/v1.0.0/
- Label your PR with the correct type (e.g., 🐛 Bug, ✨ Enhancement, 🧪
Test, etc.)
    - Provide clear and specific details in each section
-->

**Motivation:**
<!--
Explain here the context, and why you're making that change. What is the
problem you're trying to solve.
-->
In the docs we instruct the user to build from source using `go build`
this skips the `ldflags` we inject on `make install` and leaves certain
things unset (such as the version/commit).

**Modifications:**
<!--
Describe the modifications you've done from a high level. What are the
key technical decisions and why were they made?
-->
- Change instructions for building from source to use `make install`

**Result:**
<!--
*After your change, what will change.
-->
- Users building from source will see the correct version/commit when
they run `devkit version`

**Testing:**
<!--
*List testing procedures taken and useful artifacts.
-->
- Test locally