feat: standard logging  (#97)
**Motivation:**

We used to use a if verbose block for verbose logging.

**Modifications:**

- We use default zap logger for all cases now , if its a TTY or not. 
- Now we get the log level based on the top level context, if verbose is
provided, we set the logger to zap logger with development mode, i.e
debug mode , else we use production mode i.e Info. in the BeforeFunc
hook and store the logger and tracker in context. We then pass those in
the functions using the context.
- Added transaction hash and parameters for calls in contract_caller. 


**Result:**

clean up and easy to develop for maintainers

**Testing:**
<!--
*List testing procedures taken and useful artifacts.
-->
**Open questions:**

- default logger is now zap logger. Removed basic logger from TTY case. 
- Is this approach correct? What else is expected ?
- delete basic logger?