feat(devnet): Docker daemon sanity checks  (#60)
**Motivation:**

When user calls `devkit avs devnet start`, if docker desktop is not
installed, or docker desktop is not open, it fails with an ambiguous
error exit status 1.

We fix this in this PR


**Modifications:**

- First check if docker desktop is installed.If not log it , return
error with logs that user needs to install it.
- Second, if docker is installed, check if its open in the background,
else try to start it . We do for `windows` and `mac`.


**Result:**

Better UX.

**Testing:**

Manual testing .

**Open questions:**