feat: upgrade template to v0.0.10 (#138)
<!-- 
    🚨 ATTENTION! 🚨 
    
This PR template is REQUIRED. PRs not following this format will be
closed without review.
    
    Requirements:
- PR title must follow commit conventions:
https://www.conventionalcommits.org/en/v1.0.0/
- Label your PR with the correct type (e.g., 🐛 Bug, ✨ Enhancement, 🧪
Test, etc.)
    - Provide clear and specific details in each section
-->

**Motivation:**
<!--
Explain here the context, and why you're making that change. What is the
problem you're trying to solve.
-->
This PR upgrades the default template to v0.0.10 to fix an issue with
passing the correct RPC_URL through to the aggregator and executor

**Modifications:**
<!--
Describe the modifications you've done from a high level. What are the
key technical decisions and why were they made?
-->
- upgrades template version

**Result:**
<!--
*After your change, what will change.
-->
- will install the latest version on create, to upgrade existing
projects the user will run:
  
  ```devkit avs template upgrade --version v0.0.10```

**Testing:**
<!--
*List testing procedures taken and useful artifacts.
-->
- Tested locally