feat: teardown anvil container when run script is cancelled (#141)
<!-- 
    🚨 ATTENTION! 🚨 
    
This PR template is REQUIRED. PRs not following this format will be
closed without review.
    
    Requirements:
- PR title must follow commit conventions:
https://www.conventionalcommits.org/en/v1.0.0/
- Label your PR with the correct type (e.g., 🐛 Bug, ✨ Enhancement, 🧪
Test, etc.)
    - Provide clear and specific details in each section
-->

**Motivation:**
<!--
Explain here the context, and why you're making that change. What is the
problem you're trying to solve.
-->
When the user cancels after a `devnet start` we teardown the offchain
infra, but we do not kill the anvil instance.

**Modifications:**
<!--
Describe the modifications you've done from a high level. What are the
key technical decisions and why were they made?
-->
- Propagate `AvsRun` cancellation to anvil container

**Result:**
<!--
*After your change, what will change.
-->
- All containers will close following a cancellation on `devnet start`
UNLESS `--skip-avs-run` is provided

**Testing:**
<!--
*List testing procedures taken and useful artifacts.
-->
- Tested manually