feat: add a smoke test to verify release binaries (#149)
<!-- 
    🚨 ATTENTION! 🚨 
    
This PR template is REQUIRED. PRs not following this format will be
closed without review.
    
    Requirements:
- PR title must follow commit conventions:
https://www.conventionalcommits.org/en/v1.0.0/
- Label your PR with the correct type (e.g., 🐛 Bug, ✨ Enhancement, 🧪
Test, etc.)
    - Provide clear and specific details in each section
-->

**Motivation:**
<!--
Explain here the context, and why you're making that change. What is the
problem you're trying to solve.
-->
When we produce a release, we need a way to verify that the installation
and binary is working as expected.

**Modifications:**
<!--
Describe the modifications you've done from a high level. What are the
key technical decisions and why were they made?
-->
- Adds a smoke test after the release procedure to:
  - Install each of the binaries on appropriate runners
- Verify that `devkit version` returns what we expect for each of the
binaries
  - Failures will not block the release (this runs afterwards)

**Result:**
<!--
*After your change, what will change.
-->
- Additional step in the release procedure

**Testing:**
<!--
*List testing procedures taken and useful artifacts.
-->
- Tested a simplified version locally using `act`
- Will test more once this is merged

**Open questions:**
<!--
(optional) Any open questions or feedback on design desired?
-->
- Should we add e2e's that run against the binaries?