fix: ensure git profile is set before performing initial commit (#154)
<!-- 
    🚨 ATTENTION! 🚨 
    
This PR template is REQUIRED. PRs not following this format will be
closed without review.
    
    Requirements:
- PR title must follow commit conventions:
https://www.conventionalcommits.org/en/v1.0.0/
- Label your PR with the correct type (e.g., 🐛 Bug, ✨ Enhancement, 🧪
Test, etc.)
    - Provide clear and specific details in each section
-->

**Motivation:**
<!--
Explain here the context, and why you're making that change. What is the
problem you're trying to solve.
-->
We don't check if the git profile is set before performing an initial
commit. If it is not set, we should not do the commit and we should
instead warn the user.

**Modifications:**
<!--
Describe the modifications you've done from a high level. What are the
key technical decisions and why were they made?
-->
- Adds a check for the users git profile before performing initial
commit

**Result:**
<!--
*After your change, what will change.
-->
- Git does not fail because of missing profile

**Testing:**
<!--
*List testing procedures taken and useful artifacts.
-->
- Tested locally on a fresh ubuntu container