fix: pin template to v0.0.13 (#201)
<!-- 
    🚨 ATTENTION! 🚨 
    
This PR template is REQUIRED. PRs not following this format will be
closed without review.
    
    Requirements:
- PR title must follow commit conventions:
https://www.conventionalcommits.org/en/v1.0.0/
- Label your PR with the correct type (e.g., 🐛 Bug, ✨ Enhancement, 🧪
Test, etc.)
    - Provide clear and specific details in each section
-->

**Motivation:**
<!--
Explain here the context, and why you're making that change. What is the
problem you're trying to solve.
-->
As a DevKit user I expect to create from a known good version of the
`hourglass-avs-template`.

**Modifications:**
<!--
Describe the modifications you've done from a high level. What are the
key technical decisions and why were they made?
-->
- Pins template to `v0.0.13`

**Result:**
<!--
*After your change, what will change.
-->
- Create will use the pinned version as base template

**Testing:**
<!--
*List testing procedures taken and useful artifacts.
-->
- Tests have been updated