feat: upgrade commands to move to latest versions (#206)
<!-- 
    🚨 ATTENTION! 🚨 
    
This PR template is REQUIRED. PRs not following this format will be
closed without review.
    
    Requirements:
- PR title must follow commit conventions:
https://www.conventionalcommits.org/en/v1.0.0/
- Label your PR with the correct type (e.g., 🐛 Bug, ✨ Enhancement, 🧪
Test, etc.)
    - Provide clear and specific details in each section
-->

**Motivation:**
<!--
Explain here the context, and why you're making that change. What is the
problem you're trying to solve.
-->
As an AVS developer, I want a first-class command to upgrade my project
to a newer version to take advtange of the latest features with no
migration complexity.

**Modifications:**
<!--
Describe the modifications you've done from a high level. What are the
key technical decisions and why were they made?
-->
- Adds `devkit upgrade` command
- Fetches latest release from github to pull version and commit
- Avoids redundant upgrades (by comparing current to target)
- Performs upgrade, installing to `~/bin`
- Updates `devkit avs template upgrade` to default to latest template
known to devkit

**Result:**
<!--
*After your change, what will change.
-->
- Users can upgrade using `devkit upgrade` and `devkit avs template
upgrade` without defining versions

**Testing:**
<!--
*List testing procedures taken and useful artifacts.
-->
- Added unit tests to verify upgrade logic