feat: adding ecdsa keystore support  (#219)
<!-- 
    🚨 ATTENTION! 🚨 
    
This PR template is REQUIRED. PRs not following this format will be
closed without review.
    
    Requirements:
- PR title must follow commit conventions:
https://www.conventionalcommits.org/en/v1.0.0/
- Label your PR with the correct type (e.g., 🐛 Bug, ✨ Enhancement, 🧪
Test, etc.)
    - Provide clear and specific details in each section
-->

**Motivation:**
AVSs have a choice of key type to use when running on Hourglass and
different key types have different advantages. This PR introduces ECDSA
as a key option for Hourglass components and the required infrastructure
to bootstrap devkit users to take advantage of this. Through the
keystore command, users can create and read ECDSA keystores. This
feature will also be extensible to other key types in the future. Now,
based on the curve type specified for operator keys, they'll be
registered differently in the protocol's KeyRegistry contract.

**Modifications:**
- New context version
- New keystore create/read support for ECDSA 
- Refactored devkit actions to handle curve types differently for
registration
- Removed unused keystores
- New unit & integration test coverage

**Result:**
- ECDSA keystores are now supported and devkit respects the curve_type
specified from the AVS template for devkit environments.

**Testing:**
- Unit testing and manual testing by creating projects.
- Manually tested v0.0.7 -> v0.0.8 migration.

**Open questions:**
N/A