fix: pass docker-compose stdout and stderr to os stdout and stderr (#223)
<!-- 
    🚨 ATTENTION! 🚨 
    
This PR template is REQUIRED. PRs not following this format will be
closed without review.
    
    Requirements:
- PR title must follow commit conventions:
https://www.conventionalcommits.org/en/v1.0.0/
- Label your PR with the correct type (e.g., 🐛 Bug, ✨ Enhancement, 🧪
Test, etc.)
    - Provide clear and specific details in each section
-->

**Motivation:**
<!--
Explain here the context, and why you're making that change. What is the
problem you're trying to solve.
-->
If something is broken in the `docker-compose` flow, we should expose
that to the user immediately so that they can debug.

**Modifications:**
<!--
Describe the modifications you've done from a high level. What are the
key technical decisions and why were they made?
-->
- Pass `docker-compose` stdout and stderr to os stdout and stderr

**Result:**
<!--
*After your change, what will change.
-->
- User can debug problems in the anvil start-up

**Testing:**
<!--
*List testing procedures taken and useful artifacts.
-->
- Tested locally