feat: allow any context to feed core commands (#233)
<!-- 
    🚨 ATTENTION! 🚨 
    
This PR template is REQUIRED. PRs not following this format will be
closed without review.
    
    Requirements:
- PR title must follow commit conventions:
https://www.conventionalcommits.org/en/v1.0.0/
- Label your PR with the correct type (e.g., 🐛 Bug, ✨ Enhancement, 🧪
Test, etc.)
    - Provide clear and specific details in each section
-->

**Motivation:**
<!--
Explain here the context, and why you're making that change. What is the
problem you're trying to solve.
-->
As an AVS developer, I want to be able to select the chosen context and
have my selected context available to all commands. I also want to be
able to override the config set context using a `--context` flag where
appropriate

**Modifications:**
<!--
Describe the modifications you've done from a high level. What are the
key technical decisions and why were they made?
-->
- Reads selected context from config
- Override using `--context` flag

**Result:**
<!--
*After your change, what will change.
-->
- Context is controlled by config and command flags

**Testing:**
<!--
*List testing procedures taken and useful artifacts.
-->
- All tests are pulling default devnet context and passing
- Coverage already exist for context selection etc...