feat: restrict devnet only cmds (#238)
<!-- 
    🚨 ATTENTION! 🚨 
    
This PR template is REQUIRED. PRs not following this format will be
closed without review.
    
    Requirements:
- PR title must follow commit conventions:
https://www.conventionalcommits.org/en/v1.0.0/
- Label your PR with the correct type (e.g., 🐛 Bug, ✨ Enhancement, 🧪
Test, etc.)
    - Provide clear and specific details in each section
-->

**Motivation:**
<!--
Explain here the context, and why you're making that change. What is the
problem you're trying to solve.
-->
As an AVS developer, I want to be informed when I attempt to run a
command that is not meant for testnet and beyond so that I am not
surprised by any failures

**Modifications:**
<!--
Describe the modifications you've done from a high level. What are the
key technical decisions and why were they made?
-->
- Restricts various commands to `devnet` context

**Result:**
<!--
*After your change, what will change.
-->
- `devnet start/deploy-contracts`, `run` and `call` can only be called
when the selected context is `devnet`