fix: allow AVS setup to be skipped from context "flag" (#242)
<!-- 
    🚨 ATTENTION! 🚨 
    
This PR template is REQUIRED. PRs not following this format will be
closed without review.
    
    Requirements:
- PR title must follow commit conventions:
https://www.conventionalcommits.org/en/v1.0.0/
- Label your PR with the correct type (e.g., 🐛 Bug, ✨ Enhancement, 🧪
Test, etc.)
    - Provide clear and specific details in each section
-->

**Motivation:**
<!--
Explain here the context, and why you're making that change. What is the
problem you're trying to solve.
-->
As a devkit user, I need to be able to `skip-setup` on AVS deployment to
allow custom templates to handle this for me

**Modifications:**
<!--
Describe the modifications you've done from a high level. What are the
key technical decisions and why were they made?
-->
- Reverts the removal of `avs.skip-setup` context migration following
accidental removal
- Adds the same `skip-setup` procedure to `devnet start`

**Result:**
<!--
*After your change, what will change.
-->
- User can configure the `skip-setup` "flag" from their context to avoid
having to do this in `devnet start`/`deploy l1`

**Testing:**
<!--
*List testing procedures taken and useful artifacts.
-->
- Tested locally