fix: ensure we're consistently using v6.5.2 of golangci/golangci-lint-action (#244)
<!-- 
    🚨 ATTENTION! 🚨 
    
This PR template is REQUIRED. PRs not following this format will be
closed without review.
    
    Requirements:
- PR title must follow commit conventions:
https://www.conventionalcommits.org/en/v1.0.0/
- Label your PR with the correct type (e.g., 🐛 Bug, ✨ Enhancement, 🧪
Test, etc.)
    - Provide clear and specific details in each section
-->

**Motivation:**
<!--
Explain here the context, and why you're making that change. What is the
problem you're trying to solve.
-->
As a devkit dev, I want to be able to release devkit and consistently
use the same version of `golangci/golangci-lint-action` in all workflows
so that we're not identifying new issues on release.

**Modifications:**
<!--
Describe the modifications you've done from a high level. What are the
key technical decisions and why were they made?
-->
- Pins `golangci/golangci-lint-action` to `v6.5.2`

**Result:**
<!--
*After your change, what will change.
-->
- Release uses the same linting as our `golangci-lint` workflow.