feat: expose language selection to relevant template scripts (#224)
<!-- 
    🚨 ATTENTION! 🚨 
    
This PR template is REQUIRED. PRs not following this format will be
closed without review.
    
    Requirements:
- PR title must follow commit conventions:
https://www.conventionalcommits.org/en/v1.0.0/
- Label your PR with the correct type (e.g., 🐛 Bug, ✨ Enhancement, 🧪
Test, etc.)
    - Provide clear and specific details in each section
-->

**Motivation:**
<!--
Explain here the context, and why you're making that change. What is the
problem you're trying to solve.
-->
As a devkit template developer, I need to be responsive to the users
template language selection so that I can serve up the correct template
on `init`, `build` and `run`.

**Modifications:**
<!--
Describe the modifications you've done from a high level. What are the
key technical decisions and why were they made?
-->
- Restructures `templates.yaml` to more closely match new direction
(available languages in a framework)
- Store selected `language` into projects `config.yaml`
- Pass selected `language` through to template scripts on `init`,
`build` and `run`

**Result:**
<!--
*After your change, what will change.
-->
- Users template language selection can be used by template to construct
appropriate project cmd files and makeFile.

**Testing:**
<!--
*List testing procedures taken and useful artifacts.
-->
- Updated unit/integration tests and tested locally

**Open questions:**
<!--
(optional) Any open questions or feedback on design desired?
-->
- Atm, if a user provides custom `templateUrls`/`version`, we do not
check that the `language` matches to allow third party templates, in the
future, we might want to lock this down (`--experimental` flag?) to
prevent invalid `language` selection.